primary_requirements.txt
-e .[socks]
pytest>=2.8.0,<9
pytest-cov
pytest-httpbin==2.1.0
httpbin~=0.10.0
trustme
wheel
google.generativeai
pypi_simple
google.api_core.exceptions 
packaging